+++ R/gen-list.r
@@ -16,6 +16,7 @@
 #'    \item For \code{gen.list} it may have arbitrary structure (including a list).
 #'    \item For \code{gen.vector} a value (i.e., a vector of length 1) is expected.
 #'    \item For \code{gen.data.frame} a (named) vector or list is expected which describes one row of the data frame.
+#'          Use \code{list(name = val)} if \code{val} is a non-fundamental type like \code{difftime}.
 #'    \item For \code{gen.matrix} either a (named) vector/list (like \code{gen.data.frame}) or a scalar is expected. 
 #'          In the latter case we expect exactly two variables (inducing rows and columns where the order depends on \code{byrow}) within the \code{...} arguments.
 #'   }
@@ -118,7 +119,7 @@
 #' 
 #' # A data.frame containing the numbers in 2:20, the sum of their divisors
 #' # and a flag if they are "perfect" (sum of divisors equals the number)
-#' gen.data.frame(c(n, sumdiv, perfect = (n == sumdiv)), n = 2:20, 
+#' gen.data.frame(list(n, sumdiv, perfect = (n == sumdiv)), n = 2:20, 
 #'                sumdiv = sum(gen.vector(x, x = 1:(n-1), n %% x == 0)))
 #'                
 #' # A diagonal matrix with (1, ..., 5) on the diagonal